fix: use fractional-ms durations for scheduling and ACK-flush histograms
All three observe helpers (admission_compute, queue_wait, pending_queue_age)
accepted integer ms types (u128/u64) and cast to f64 at observe time.
Admission compute buckets start at 0.001 ms and queue_wait/pending_queue_age
at 0.01 ms; the integer conversion truncated sub-1ms durations to 0, making
the fast-path distribution invisible. Same issue in ACK write-queue and
socket-write observations in shared_tcp_endpoint (bucket floor 0.01 ms).

Fix: change all three observe signatures to f64, update every call site to
use as_secs_f64() * 1000.0, and inline the secs_f64 computation directly at
the two shared_tcp_endpoint ACK observe call sites.

Refs: DYN-3401

+++ lib/kv-router/src/scheduling/metrics.rs
@@ -140,16 +140,16 @@ impl SchedulingMetrics {
         SCHEDULING_METRICS.get_or_init(|| Self::new().expect("scheduling metrics"))
     }
 
-    pub fn observe_admission_compute(&self, worker_type: &str, ms: u128) {
+    pub fn observe_admission_compute(&self, worker_type: &str, ms: f64) {
         self.admission_compute_ms
             .with_label_values(&[worker_type])
-            .observe(ms as f64);
+            .observe(ms);
     }
 
-    pub fn observe_queue_wait(&self, worker_type: &str, ms: u64) {
+    pub fn observe_queue_wait(&self, worker_type: &str, ms: f64) {
         self.queue_wait_ms
             .with_label_values(&[worker_type])
-            .observe(ms as f64);
+            .observe(ms);
     }
 
     pub fn observe_actor_mailbox_wait(&self, worker_type: &str, command: &str, ms: f64) {
@@ -170,10 +170,10 @@ impl SchedulingMetrics {
             .observe(ms);
     }
 
-    pub fn observe_pending_queue_age(&self, worker_type: &str, ms: u64) {
+    pub fn observe_pending_queue_age(&self, worker_type: &str, ms: f64) {
         self.pending_queue_age_ms
             .with_label_values(&[worker_type])
-            .observe(ms as f64);
+            .observe(ms);
     }
 
     pub fn observe_worker_state_update_to_scheduler(&self, event: &str, ms: f64) {
